memory/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
supabase==2.3.0
python-dotenv==1.0.0
rich==13.7.0
